Move vault test code to files where it is used
